+++ databricks-sdk-java/src/main/java/com/databricks/sdk/core/retry/RetryStrategy.java
@@ -4,7 +4,7 @@ import com.databricks.sdk.core.DatabricksError;
 import com.databricks.sdk.core.http.Response;
 
 public interface RetryStrategy {
-    boolean isRetriable(Response response, Exception e);
+  boolean isRetriable(Response response, Exception e);
 
-    DatabricksError getError();
+  DatabricksError getError();
 }
