+++ databricks-sdk-java/src/main/java/com/databricks/sdk/core/retry/RetryStrategyPicker.java
@@ -3,5 +3,5 @@ package com.databricks.sdk.core.retry;
 import com.databricks.sdk.core.http.Request;
 
 public interface RetryStrategyPicker {
-    RetryStrategy getRetryStrategy(Request request);
+  RetryStrategy getRetryStrategy(Request request);
 }
